feat: rename job and add psup period to config
